unify retry logic to pkg
